refactored variable initialization, test fix
