+++ front-end/src/Views/UserProfilePage/components/CVApprovalButton/CVApprovalButton.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import {LinkContainer} from "react-router-bootstrap";
 import {Button} from "react-bootstrap";
-import {staffTypes} from "constants/routes";
+import {staffTypes} from "constants/staffTypes";
 
-const CVApprovalButton = ({userType}) => (
-    userType.type === "Staff" && userType.data.group_type === staffTypes.CV ? (
+const CVApprovalButton = ({user}) => (
+    user.type === "Staff" && user.data.group_type === staffTypes.CV ? (
         <LinkContainer to="/cvApproval">
             <Button variant="primary" className="ml-3">
                 Zobacz CV do akceptacji
