+++ front-end/src/constants/paths.js
@@ -1,0 +1,22 @@
+const paths = {
+    DASHBOARD: "/",
+    CV_CREATOR: "/cvEditor",
+    CV_EDITOR: "/cvEditor/:id",
+    REGISTER: "/newAccount/",
+    REGISTER_ADMIN: "/newAccount/:role(staff)",
+    FOOTER: "/footer",
+    LOGIN: "/login",
+    USER: "/user",
+    OFFER_FORM: "/offerForm/:id?",
+    JOB_OFFERS: "/jobOffers",
+    CV_APPROVAL: "/cvApproval",
+    MY_OFFERS: "/myOffers",
+    CONTACT_PAGE: "/contact",
+    BLOG_POST: "/blog/blogpost/:id",
+    BLOG_PAGE: "/blog",
+    BLOG_FORM: "/blog/newPost",
+    BLOG_EDIT: "/blog/newPost/:id",
+    JOB_OFFER_DETAILS: "/jobOffers/:id",
+    MY_CVS: "/myCVs",
+};
+export {paths};
