+++ front-end/src/constants/staffTypes.js
@@ -1,0 +1,7 @@
+export const staffTypes = {
+    VERIFICATION: "staff_verification",
+    CV: "staff_cv",
+    JOBS: "staff_jobs",
+    BLOG_CREATOR: "staff_blog_creator",
+    BLOG_MODERATOR: "staff_blog_moderator",
+};
