+++ front-end/src/constants/userTypes.js
@@ -1,0 +1,5 @@
+export const userTypes = {
+    STANDARD: "Standard",
+    STAFF: "Staff",
+    EMPLOYER: "Employer",
+};
